+++ src/ui/utils/getEligibleCollections.js
@@ -1,3 +1,5 @@
+import { DEFAULT_COLLECTION_YEAR } from '__configs/systemConfig';
+
 const getEligibleCollections = (
     timeStamp,
     isDefaultCollection,
@@ -12,12 +14,21 @@ const getEligibleCollections = (
             return false;
         }
 
-        if (!isDefaultCollection && item.year !== year) {
+        if (excludeCollectionIds.includes(item.id)) {
             return false;
         }
 
-        if (excludeCollectionIds.includes(item.id)) {
-            return false;
+        if (!isDefaultCollection) {
+            if (item.year === DEFAULT_COLLECTION_YEAR) {
+                return true;
+            }
+            if (item.year !== year) {
+                return false;
+            }
+        } else {
+            if (item.year !== year) {
+                return false;
+            }
         }
         return true;
     });