+++ CHANGELOG.md
@@ -7,17 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.1] - 2026-07-28
+
 ### Added
 
-- List of features/files added
+- Fullscreen handler added.
+- Added search for Audio
 
 ### Changed
 
-- List of changes
+- All service files with missing consistent response code are changed.
+- Search service changed to add audio support
 
 ### Fixed
 
-- List of fixes made
+- Fixed inconsistencies
+- Error in audio and image delete operations
+- Fixed first time loading of media details problem
 
 ## [0.2.0] - 2026-07-26